Add callback.go to unify Callback implementation
